test: jest 설정 및 예제 작성
