+++ frontend/babel.config.js
@@ -1,0 +1,9 @@
+const config = {
+  "presets": [["@babel/preset-env", {
+    "targets": {
+      "node": "current"
+    }
+  }], "@babel/preset-react", "@babel/preset-typescript"]
+};
+
+export default config;
