--- frontend/babel.config.json
@@ -1,14 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "targets": {
-          "node": "current"
-        }
-      }
-    ],
-    "@babel/preset-react",
-    "@babel/preset-typescript"
-  ]
-}
