+++ frontend/package.json
@@ -7,6 +7,7 @@
     "dev": "vite",
     "build": "tsc && vite build",
     "preview": "vite preview",
+    "prebuild": "npm run test",
     "test": "jest"
   },
   "dependencies": {
@@ -30,7 +31,10 @@
     "@types/react-dom": "^18.0.8",
     "@types/styled-components": "^5.1.26",
     "@vitejs/plugin-react": "^2.2.0",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
+    "jest-transform-stub": "^2.0.0",
     "ts-jest": "^29.0.3",
     "typescript": "^4.6.4",
     "vite": "^3.2.3"
