+++ frontend/src/assets/icons/index.ts
@@ -2,6 +2,6 @@ import SliderLeft from "./SliderLeft.svg";
 import SliderRight from "./SliderRight.svg";
 import FilterButton from "./FilterButton.svg";
 import RedDelete from "./RedDelete.svg";
-import Delete from "./Delete.png";
+import Delete from "./RedDelete.svg";
 
 export { SliderLeft, SliderRight, FilterButton, RedDelete, Delete };
