+++ frontend/src/main.tsx
@@ -2,17 +2,14 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
 
-import {BrowserRouter} from "react-router-dom";
 import {RecoilRoot} from 'recoil';
 import {GlobalStyle} from "./styles/GlobalStyle";
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-    <React.StrictMode>
-        <GlobalStyle/>
-        <RecoilRoot>
-            <BrowserRouter>
-                <App/>
-            </BrowserRouter>
-        </RecoilRoot>
-    </React.StrictMode>
+  <React.StrictMode>
+    <GlobalStyle/>
+    <RecoilRoot>
+      <App/>
+    </RecoilRoot>
+  </React.StrictMode>
 )
