+++ frontend/test/setup.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import "@testing-library/jest-dom";
